requirements.txt
requests
pandas
openai
apache-airflow
langchain
bs4
simplegmail
google-search-results
# -e .